requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
pandas>=2.1.3
openpyxl>=3.1.2
celery>=5.3.6
redis>=5.0.1
python-dotenv>=1.0.0
gunicorn>=21.2.0